Rename hash file
