+++ src/amulet/pybind11_extensions/hash.hpp
