+++ tests/_test_hash.py.cpp
@@ -1,6 +1,6 @@
 #include <pybind11/pybind11.h>
 
-#include <amulet/pybind11_extensions/hash.py.hpp>
+#include <amulet/pybind11_extensions/hash.hpp>
 
 namespace py = pybind11;
 namespace pyext = Amulet::pybind11_extensions;
